requirements.txt
Flask==2.3.3
openai==1.12.0
python-dotenv==1.0.0
markdown==3.5.1
Werkzeug==2.3.7
watchdog==3.0.0